LPS-95113 Check before deletion whether the Layout has already been deleted as a child of another Layout

+++ modules/apps/web-experience/export-import/export-import-service/src/main/java/com/liferay/exportimport/controller/LayoutImportController.java
@@ -696,7 +696,11 @@ public class LayoutImportController implements ImportController {
 					sourcePrototypeLayoutUuid, layoutSetPrototype.getGroupId(),
 					true);
 
-				if (sourcePrototypeLayout == null) {
+				if ((sourcePrototypeLayout == null) &&
+					(_layoutLocalService.fetchLayout(
+						layout.getUuid(), layout.getGroupId(),
+						layout.isPrivateLayout()) != null)) {
+
 					_layoutLocalService.deleteLayout(
 						layout, false, serviceContext);
 				}
